requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.11.7
python-multipart==0.0.6 
websockets==11.0.3
pillow==11.3.0